* USE ONLY PROPER AMERICAN ENGLISH STANDARD WORDS!!!!!!
signall vs. signal .. cancell vs. cancel ! My god man!

There are WAY more (British vs. American issues and use
is TOTALLY INCONSISTENT within the code base), but these
are, to me, the most egregious and wrong-looking ones.
The ones that are acceptable like "standardisation" use
the American "standardization" - none of it makes sense.

There are dozens if not hundreds of cases of mixed use,
often in the same file, sometimes in the same paragraph!

I've really only scratched the surface of it with this PR.

Just looking at these, though, is giving me eye cancer,
dissolving my soul, accelerating my descent into insanity.

I do think we should adopt American English as THE standard,
actually, for both documentation and code, American English
has always clearly been the standard language of the Internet.

Over at the supposedly "international" ISO, you can see the
"The International Organization for Standardization" and NOT
the "The British and Welsh Organisation for Standardisation".

Also, all the seed words we use are ONLY in American English.
We got "color" & "labor" & "humor" there. We got "truck" also,
but not a single "lorry" to be found.  If we are going to keep
using a mixed language base, would we have to change these too?

If we keep mixing, people will just think we are Canadians! ***
(*** Of course, being dead is preferred to being Canadian.)

[FYI: Merging this would increase my productivity by 20%.]