* Forked goleveldb
Virgin upstream goleveldb v1.0.1-0.20200815110645-5c35d600f0ca.

The only changes I've made to anything here are just what is
absolutely necessary for proper inclusion in the pktd source,
such as the module name and simple go.mod harmonization, to
avoid pulling in different versions of external dependencies,
removing the .git metadata, etc.

At first, I tried doing this as a git submodule, and again as
a Go vendored module dependency, and a kind of hybrid (using
the kovetskiy/manul tool), using the Golang "bundle" utility
from the Golang toolchain utilites, but each one adds different
issues and makes things much more complicated when compared to
simple inclusion of a forked version of the golevedb module.

It might be best to not include this as pkt-cash/pktd/goleveldb,
but let it live as pkt-cash/goleveldb, so you avoid "polluting"
the pktd commit logs with goleveld-related changes and rebases.

It will also be easier if a new goleveldb version is properly
released, since the package is being actively maintained again.