As mentioned/promised in issue #82, here's the PR.
This will head off any potential problem with the
not-out-yet-but-soon-to-be-released Golang 1.16.

It was recommended to me, that when updating x/net,
x/sys be updated as well, so that was also done.

gRPC and genproto also bumped - the purpose here is to
eliminate the protobuf dependency we currently have.

I did a light review of the changes and some initial
testing, all of which seems sane and was successful.

--

If you adopt PR#78/#79, and we begin to use our forked
goleveldb, then x/net and x/sys should be updated to be
the same versions there as well (FYI my initial tests
showed no issues), but it makes sense we do it, merely
to avoid pulling in multiple versions of the external
modules when we can do so; and if we're using own own,
then certainly we should do that.

Let me know - if you do adopt those PR's, I can then
either resubmit PR after rebase with these inclusions,
or make changes to PR #78, whatever is easier for you.

--