Massive improvement in bbolt/walletdb speed
First, we take the size of the walletdb and set our
initial mmap size to it's current size * 2 - this
will reduces the amount of remaping that needs to
be done at doubling and also reduces memory
fragmentation - actual results is less memory usage,
though higher usage initially.

Also enable the NoFreelistSync option to not sync
the freelist to disk, which has a massive improvement
on write performance.

Combined, tests on a Gridnode show a 250% gain.