Align with Satoshi Bitcoin Core: Alert removal.
* Removed from Bitcoin Core since 2016.

The alert system was hastily implemented by
Satoshi Nakamoto after the value overflow
incident on August 15, 2010. Satoshi never
actually used this system; it remained dormant
until the February 20, 2012 protocol change, for
which an alert was issued on February 18.

In 2016, the alert system was removoed because
of the possibility of privileged users sending
political alert messages, and because of the
possibility of the alert key having been taken
from Mark Karpelès by Japanese police in 2014.

The alert key was scheduled to be released to
the public in May 2017, and was eventually
published in July 2018, allowing unlimited use.