Overhaul of the coverage analysis framework:
Rework the original coverage reporting from the
shell scripts scattered throughout the repository
into a single script which acts as a driver for
both gocov and gocov-html, producing reports in
JSON, text, and HTML formats. I will extend this
to support Coveralls.io and GitHub hooks later.

The following is actual output from this release:

* https://pktdist.gridfinity.com/.gocov/gocov_report_pktd.html
* https://pktdist.gridfinity.com/.gocov/gocov_report_goleveldb.html

I have also modified the .gitignore, which you
may decide against, for example, if you want to
commit and include these code coverage reports
to the repo, like the Conformal people used to.