Do not run the test suite in parallel by default
This PR changes the default for most tests from
running in parallel to not - the rationale is
that we've been plagued by what are essentially
"false positives" from Travis CI, most of which
are caused by improper test parallelization.

This let's the test suite run serially without
the need to pass flags (-parallel=1) to `go test`
and will eliminate most of these false positives
while still having the test suite be effective.

Many of the tests that were failing due to data
race issues do so at very low rates - see #128:
https://github.com/pkt-cash/pktd/issues/128 for
just one example. The Golang race detector in
combination with explicitly requesting parallel
tests is going to be the only real effective
way to get useful results when a test is failing
due to proper way we should be running these
parallelized tests and with a high -count to
ensure that the data is useful.

In short, all the random failures from CI/CD
when changing a comment or documentation are
less than useful, and properly running the
test suite in parallel requires more than one
iteration of each test, and parallelized test
failures without the use of the race detector
do not produce any useful feedback as to what
is causing the failing test to actually fail,
so there is no real downside to this approach.

I will submit a future PR that does do proper
parallel testing, by using proper parallelized
tests - in Golang, the use of t.Parallel():
"signals that this test is to be run in
parallel with (and only with) other parallel
tests. When a test is run multiple times due to
use of -test.count or -test.cpu, multiple
instances of a single test never run in parallel
with each other." - so the "random failures"
we are seeing are the result of other concurrent
tests being run and not necessarily the function
under test: if these are genuine races, the race
detector should be used to give useful feedback,
rather than the useless CI feedback we get now.

I'm already working on these new tests that are
properly and explicitly parallelized, so actually
useful for detecting problems within the function
or appropriate to the condition being tested,
rather than just the random failures from the
interactions with other running tests we have now.

Those will be sent as a separate pull request,
once they are ready - and properly tested, with
their own tests!