Increase rebroadcast interval and ounbounds
The intention of this function is to rebroadcast
transactions *between* blocks, however, with the
increased speed of the PKT chain, the default for
the rebroadcast interval was the chain block time
itself, rendering it less than effective.  With
this patch, we increase the rebroadcast timing to
match the PKT chain paramters, and also increase
the number of attempted outbound connections from
8 to 12, which gives moderate improvements under
simulated test conditions meant to approximate the
PKT mainnet.